[#189] Split tokenBuffer into main and Writer sections
